+++ module/arm5e.js
@@ -44,7 +44,13 @@ import {
 import { AbilitySchema } from "./schemas/abilitySchema.js";
 import { BookSchema } from "./schemas/bookSchema.js";
 import { DiaryEntrySchema } from "./schemas/diarySchema.js";
-import { ItemSchema, VirtueFlawSchema, VisSchema } from "./schemas/minorItemsSchemas.js";
+import {
+  ItemSchema,
+  PersonalityTraitSchema,
+  ReputationSchema,
+  VirtueFlawSchema,
+  VisSchema
+} from "./schemas/minorItemsSchemas.js";
 import { LabSchema } from "./schemas/labSchema.js";
 
 Hooks.once("init", async function() {
@@ -363,6 +369,8 @@ function setSystemDatamodels() {
   CONFIG.Item.systemDataModels["spell"] = SpellSchema;
   CONFIG.Item.systemDataModels["laboratoryText"] = LabTextSchema;
   CONFIG.Item.systemDataModels["diaryEntry"] = DiaryEntrySchema;
+  CONFIG.Item.systemDataModels["personality"] = PersonalityTraitSchema;
+  CONFIG.Item.systemDataModels["reputation"] = ReputationSchema;
   //Actors
   CONFIG.Actor.systemDataModels["laboratory"] = LabSchema;
 }