+++ module/schemas/minorItemsSchemas.js
@@ -1,7 +1,13 @@
 // import DataModel from "common/abstract/data.mjs";
 import { ARM5E } from "../config.js";
 import { log } from "../tools.js";
-import { boolOption, convertToInteger, convertToNumber, itemBase } from "./commonSchemas.js";
+import {
+  boolOption,
+  convertToInteger,
+  convertToNumber,
+  itemBase,
+  XpField
+} from "./commonSchemas.js";
 const fields = foundry.data.fields;
 
 export const possibleReputationTypes = Object.keys(ARM5E.reputations);
@@ -44,6 +50,7 @@ export class VirtueFlawSchema extends foundry.abstract.DataModel {
     if (data.type?.value) {
       data.type = data.type.value;
     }
+    return data;
   }
 
   static migrate(itemData) {
@@ -132,6 +139,8 @@ export class VisSchema extends foundry.abstract.DataModel {
     if (data.art?.value) {
       data.art = data.art.value;
     }
+
+    return data;
   }
 
   static migrate(itemData) {
@@ -172,8 +181,22 @@ export class ReputationSchema extends foundry.abstract.DataModel {
     };
   }
 
+  static migrateData(data) {
+    // console.log(`MigrateData Reputation: ${JSON.stringify(data)}`);
+    if (data.points != undefined) {
+      data.xp = (5 * (data.points * (data.points + 1))) / 2;
+      delete data.points;
+    }
+    return data;
+  }
+
   static migrate(data) {
-    return {};
+    // console.log(`Migrate Reputation: ${JSON.stringify(data)}`);
+    let update = {};
+    update["system.-=points"] = null;
+    update["system.-=notes"] = null;
+    update["system.xp"] = data.system.xp;
+    return update;
   }
 }
 
@@ -192,6 +215,11 @@ export class PersonalityTraitSchema extends foundry.abstract.DataModel {
   static migrate(data) {
     return {};
   }
+
+  static migrateData(data) {
+    console.log(`MigrateData Personality trait: ${JSON.stringify(data)}`);
+    return data;
+  }
 }
 
 export class MySchema extends foundry.abstract.DataModel {
@@ -200,6 +228,10 @@ export class MySchema extends foundry.abstract.DataModel {
 
   static defineSchema() {
     return { ...itemBase() };
+  }
+
+  static migrateData(data) {
+    return data;
   }
 
   static migrate(data) {